Lots of little changes to bring OggConvert mostly inline with OggConvert++ (except for the new media checker)
Changes:

* ocv_gtk.py split into ocv_main.py and ocv_in_progress.py
* ocv_gst.py split into ocv_transcoder.py and ocv_media_checker.py
* Lots of tidying up of messy functions in ocv_main
* The media checker is now callback-based. It still blocks the UI, but 
at least it doesn't *look* like it blocks the UI any more ;-). It's 
still using the discoverer at the moment, but that will change
* Brought A. Bram Neijt's OggConvert icon into the source tree and use 
it as the main icon for the app.
* The URL in the About dialogue is now clickable, opening a new browser 
tab. I know how important this is.
* Minor change to the way the time remaining is displayed.
* Audio options are greyed-out when a video-only file is selected, and 
vice-versa.
* Use the extension ".ogv" for Ogg video files (both Theora and Dirac), 
in accordance with the new Xiph.org guidelines. Fixes bug #160054.

Still to do:

* Port the OggConvert++ media checker to Python
* Get the transcoder to send out "paused" and "playing" signals which 
the GUI can listen for.
* Investigate whether we can support Gnome-VFS for input/output without 
depending on it (which I don't want to do)
* Bring in all the translations for Launchpad.
* Release 0.3.0
* Party like it's 1999.