New media checker working, quite nicely I think. Slightly faster than the old discoverer, and hopefully with fewer problems with e.g. flv files.
The transcoder is now connected to the GUI the "proper" way, i.e. using 
gobject callbacks. The way the timing is done is now much more accurate. 
Shame the method of predicting time remaining (assuming completely 
linear encoding progress) still sucks, but there we go.

Removed the "stall warning" dialogue, since (a) it was really badly 
worded, (b) all the translators had a hard time with it, and (c) 99% of 
the time it was a false alarm anyway.

Nearly ready for a 0.3.0 release I think...