Updated changelog, readme, and todo files for release. Updated to latest Launchpad translations, added Dutch.
Set the minimum window width (back) to 360px, because I think it looks 
better this way.

Don't display time remaining for the first three seconds, as we get 
crazy numbers during this time (as GStreamer often likes to report a 
negative stream position which messes things up).

And that's it: this is 0.3.0.